Change to numbered menu
Menu now extensable and allows the user to select
an integer instead of pressing the down arrow a
number of times.
Would like to generalize and refactor more but it
is time to move onto the heart of the program.

Next,
Change "Edit" options to "Show"
Create simple input files and begin developing the core Calculate algorithm, then work out from there.