Add more catalogue tests and notes
Deciding if I should have
<string name, Product>
    pairs in which the name is repeated,
<string, Amount> pairs, or
<string, decimal> pairs

Just use the pair in the dictionary to represent
what I've been calling a product.

I'll continue with the current experiment for now.

ExampleCatalogueAsDictOfProducts.json:
{
  "Pineapple": {
    "RegularPrice": {
      "Symbol": "$",
      "Amount": 1.08
    },
    "Name": "Pineapple"
  },
  "Pepperoni": {
    "RegularPrice": {
      "Symbol": "$",
      "Amount": 5.73
    },
    "Name": "Pepperoni"
  }
}

I can see that this will work but it is a bit
over done. Having the name as a property and the
key of each entry is redundant. Having the Symbol
repeated for each entry is redundant.

I'm am considering symplifying things knowing I
can add the product object back in if needed.

All that is needed now is:

{
  "Pineapple": 1.08,
  "Pepperoni": 5.73
}

But, using the json serializers means I don't have
to do the serialization manually. The output to
the screen or a saved file for users can be
different. Will continue the experiment with a
new ToString() method. The dictionary of the
catalogue would mostly likely be data stored
behind a database engine.

What is the process of reading in the basket?
It will result in a sorted dictionary of product
name as the key and the quantity as the value.

The calculation will begin with retrieving the
regular price for each of the product names in
the sorted dictionary. This will require the name
as  the key and then pulling out the
product.regularprice.amount and appending that to
 the value in the dictionary
then applay promitions by item and quantity to
get the effective price for the product

Then the total can be found by totalling the
effective prices. These results can be achieve by
a more functional style. Would the process be
easier if the amount is pulled from a name: amount
paring in the dicionary. I don't see there being a
big difference.

Next, make a new diplay method.